+++ lib/table_transform.rb
@@ -1,3 +1,4 @@
 require 'table_transform/version'
 require 'table_transform/table'
+require 'table_transform/formula_helper'
 require 'table_transform/excel_creator'