+++ lib/table_transform/table.rb
@@ -8,6 +8,8 @@ module TableTransform
   end
 
   class Table
+    attr_reader :formulas
+
     def self.create_from_file(file_name, sep = ',')
       rows = CSV.read(file_name, { :col_sep => sep })
       raise "'#{file_name}' contains no data" if rows.empty?
@@ -30,6 +32,7 @@ module TableTransform
       header = @data_rows.shift
       @column_indexes = create_column_name_binding(header)
       @metadata = header.zip( Array.new(header.size){{}} ).to_h
+      @formulas = {}
 
       validate_header_uniqueness(header)
       validate_column_size
@@ -48,6 +51,12 @@ module TableTransform
     # Returns meta data as Hash with header name as key
     def metadata
       @metadata.clone
+    end
+
+    def add_column_formula(column, formula, metadata = {})
+      add_column(column, metadata){nil}
+      @formulas[column] = formula
+      self # self chaining
     end
 
     def << (hash_values)
@@ -83,6 +92,7 @@ module TableTransform
       selected_cols = @column_indexes.values_at(*header)
       t = Table.new( @data_rows.inject([header]) {|res, row| (res << row.values_at(*selected_cols))} )
       t.metadata = t.metadata.keys.zip(@metadata.values_at(*header)).to_h
+      t.formulas = header.zip(@formulas.values_at(*header)).to_h
       t
     end
 
@@ -105,6 +115,7 @@ module TableTransform
     end
 
     def change_column(name)
+      raise "Column with formula('#{name}') cannot be changed" if @formulas[name]
       index = Util::get_col_index(name, @column_indexes)
       @data_rows.each{|r|
         r[index] = yield Row.new(@column_indexes, r)
@@ -115,7 +126,10 @@ module TableTransform
 
     def delete_column(*names)
       validate_column_names(*names)
-      names.each{|n| @metadata.delete(n)}
+      names.each{|n|
+        @metadata.delete(n)
+        @formulas.delete(n)
+      }
 
       selected_cols = @column_indexes.values_at(*@metadata.keys)
       @data_rows.map!{|row| row.values_at(*selected_cols)}
@@ -124,6 +138,8 @@ module TableTransform
       self
     end
 
+    # Table row
+    # Columns within row can be referenced by name, e.g. row['name']
     class Row
 
       def initialize(cols, row)
@@ -140,6 +156,7 @@ module TableTransform
 
     end
 
+    # Cell within Table::Row
     class Cell < String
       # @returns true if this cell includes any of the given values in list
       def include_any?(list)
@@ -149,6 +166,7 @@ module TableTransform
 
     protected
       attr_writer :metadata
+      attr_writer :formulas
 
     private
       def create_column_name_binding(header_row)