+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 TableTransform is still in pre-release state. This means that its APIs and behavior are subject to breaking changes without deprecation notices. Until 1.0, version numbers will follow a [Semver][]-ish `0.y.z` format, where `y` is incremented when new features or breaking changes are introduced, and `z` is incremented for lesser changes or bug fixes.
 
 ## [Unreleased]
+* Supports formulas in columns
+* Helper functions to create formulas
+* Column width estimates format size in calculation 
 
 ## [0.3.0][] (2016-02-29)
 * Added format capability for columns when published as Excel